Update espercli.md
This is Keith. Not many proposed changes in this one. One question is the mismatch between the Python SDK and the CLI regarding Python version. The Python SDK recommends 3.7 instead of 2.7 - which will work. The assumption is this is so that developer can use both the CLI and the SDK with the one python installation. However the CLI above states 3.6. So should the SDK be 3.6, or should the CLI be 3.7?