Update gettingstarted.md
Szot here again. I restructured this a bit to make it more complete step-by-step for what we want developers to do. Please reach out to me if you think this should be done differently. I am making this about the API-level and Android Studio as much as I can. There are links that you should drop in that I did not, I think they will be obvious. Thanks!