Update emulator.md
What is the plan to provide documentation on the Virtual Emulator? If we don't have it available yet, this page should not be up. Let's rally for Sprint 20 to see where we are at and what we have, and drop this page back in. It does not look good to have this kind of page in there right now in public view, so delete it until we have content.